+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased](https://github.com/USGS-WiM/stnweb2/tree/dev)
 
 ### Added
- -
+
+ - Ability to create HWMs
  
 ### Changed
  -